iot-core-lab/simulator/requirements.txt
awsiotsdk>=1.21.0
paho-mqtt>=2.0.0
boto3>=1.28.0
